@@ -7,10 +7,15 @@
 
     "@testing-library/react": "^11.2.7",
 
+
+    "@material-ui/core": "^4.11.4",
+    "@material-ui/icons": "^4.11.2",
     "@testing-library/user-event": "^13.1.9",
+
+
     "axios": "^0.21.1",
     "bootstrap": "^4.6.0",
-    "multiselect-react-dropdown": "^1.6.11",
+    "multiselect-react-dropdown": "^1.6.13",
     "react": "^17.0.2",
     "react-awesome-loaders": "^0.1.37",
     "react-dom": "^17.0.2",
@@ -18,6 +23,8 @@
     "react-icons": "^4.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "reactstrap": "^8.9.0",
+    "styled-components": "^5.3.0",
     "react-scroll": "^1.8.2",
     "reactstrap": "^8.9.0"
   },
@@ -44,7 +51,7 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^7.26.0",
+    "eslint": "^7.27.0",
     "eslint-plugin-react": "^7.23.2",
     "prettier": "2.3.0"
   }
