@@ -13,7 +13,7 @@ import matplotlib.pyplot as plt
 import seaborn as sns
 
 # we want our plots to appear inside the notebook
-get_ipython().run_line_magic('matplotlib', 'inline')
+get_ipython().run_line_magic("matplotlib", "inline")
 
 # Models from Scikit-Learn
 from sklearn.linear_model import LogisticRegression
@@ -34,7 +34,7 @@ from sklearn.metrics import plot_roc_curve
 # In[2]:
 
 
-hrtds = pd.read_csv('heart-disease.csv')
+hrtds = pd.read_csv("heart-disease.csv")
 hrtds.head()
 
 
@@ -85,66 +85,65 @@ hrtds.dtypes
 # In[10]:
 
 
-hrtds.plot.hist(figsize=(10, 30), subplots=True);
+hrtds.plot.hist(figsize=(10, 30), subplots=True)
 
 
 # In[11]:
 
 
-sns.pairplot(hrtds,height=30);
+sns.pairplot(hrtds, height=30)
 
 
 # In[ ]:
 
 
-
-
-
 # In[12]:
 
 
-hrtds["target"].value_counts().plot( x = 'Age', y= 'count',kind="bar", color=["salmon", "lightblue"])
+hrtds["target"].value_counts().plot(
+    x="Age", y="count", kind="bar", color=["salmon", "lightblue"]
+)
 
 
 # In[13]:
 
 
-pd.crosstab(hrtds.target,hrtds.sex)
+pd.crosstab(hrtds.target, hrtds.sex)
 
 
 # In[14]:
 
 
-pd.crosstab(hrtds.target,hrtds.sex).plot(kind="bar",
-                                   figsize=(10,6), 
-                                   color=["lightblue", "salmon"])
+pd.crosstab(hrtds.target, hrtds.sex).plot(
+    kind="bar", figsize=(10, 6), color=["lightblue", "salmon"]
+)
 plt.legend(["Female", "Male"]),
 
 
-plt.ylabel(" ");
+plt.ylabel(" ")
 
 
 # In[15]:
 
 
 # Create another figure
-plt.figure(figsize=(10,6))
+plt.figure(figsize=(10, 6))
 
 # Start with positve examples
-plt.scatter(hrtds.age[hrtds.target==1], 
-            hrtds.thalach[hrtds.target==1], 
-            c="salmon") # define it as a scatter figure
+plt.scatter(
+    hrtds.age[hrtds.target == 1], hrtds.thalach[hrtds.target == 1], c="salmon"
+)  # define it as a scatter figure
 
 # Now for negative examples, we want them on the same plot, so we call plt again
-plt.scatter(hrtds.age[hrtds.target==0], 
-            hrtds.thalach[hrtds.target==0],
-            c="lightblue") # axis always come as (x, y)
+plt.scatter(
+    hrtds.age[hrtds.target == 0], hrtds.thalach[hrtds.target == 0], c="lightblue"
+)  # axis always come as (x, y)
 
 # Add some helpful info
 plt.title("Heart Disease in function of Age and Max Heart Rate")
 
 plt.legend(["Disease", "No Disease"])
-plt.ylabel("Max Heart Rate");
+plt.ylabel("Max Heart Rate")
 
 
 # Model Training
@@ -180,9 +179,9 @@ y
 np.random.seed(42)
 
 # Split into train & test set
-X_train, X_test, y_train, y_test = train_test_split(X, # independent variables 
-                                                    y, # dependent variable
-                                                    test_size = 0.2) # percentage of data to use for test set
+X_train, X_test, y_train, y_test = train_test_split(
+    X, y, test_size=0.2  # independent variables  # dependent variable
+)  # percentage of data to use for test set
 
 
 # In[20]:
@@ -234,7 +233,7 @@ ypreds
 # In[27]:
 
 
-model.score(X_test,y_test)#evaluating model
+model.score(X_test, y_test)  # evaluating model
 
 
 # In[28]:
@@ -253,17 +252,15 @@ X_test.loc[104]
 
 
 pred = model.predict(X_test.loc[:193])
-pred #the last element of this array is our prediction for patient no. 193)
+pred  # the last element of this array is our prediction for patient no. 193)
 
 
 # In[40]:
 
 
-hrtds.loc[193] #we can check that with our previous data that our prediction is correct.
+hrtds.loc[
+    193
+]  # we can check that with our previous data that our prediction is correct.
 
 
 # In[ ]:
-
-
-
-
