@@ -1,5 +1,6 @@
 from django.contrib import admin
-from .models import PatientRegister,DoctorRegister,Emergency,Appointment
+from .models import PatientRegister, DoctorRegister, Emergency, Appointment
+
 # Register your models here.
 admin.site.register(PatientRegister)
 admin.site.register(DoctorRegister)
