@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class LifemedeasyConfig(AppConfig):
-    name = 'lifeMEDeasy'
+    name = "lifeMEDeasy"
