@@ -14,17 +14,24 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path,include
-from django.views.generic import TemplateView
-from lifeMEDeasy.views import PatientRegisterView,DoctorRegisterView,DPLogin,AmbulanceView,PatientResource,DoctorResource,AppointmentView
+from django.urls import path, include
+from lifeMEDeasy.views import (
+    PatientRegisterView,
+    DoctorRegisterView,
+    DPLogin,
+    AmbulanceView,
+    PatientResource,
+    DoctorResource,
+    AppointmentView,
+)
+
 urlpatterns = [
-    path('patients',include(PatientResource.urls())),
-    path('doctors',include(DoctorResource.urls())),
-    path('admin/', admin.site.urls),
-    path('patient-register',PatientRegisterView),
-    path('doctor-register',DoctorRegisterView),
-    path('login',DPLogin),
-    path('emergency',AmbulanceView),
-    path('appointment',AppointmentView)
-    
+    path("patients", include(PatientResource.urls())),
+    path("doctors", include(DoctorResource.urls())),
+    path("admin/", admin.site.urls),
+    path("patient-register", PatientRegisterView),
+    path("doctor-register", DoctorRegisterView),
+    path("login", DPLogin),
+    path("emergency", AmbulanceView),
+    path("appointment", AppointmentView),
 ]
