@@ -1,4 +1,4 @@
-import React, { Suspense } from 'react';
+import React, { Suspense, useState } from 'react';
 import NavbarItem from './components/NavbarItem';
 import Footer from './components/Footer';
 import { BrowserRouter } from 'react-router-dom'
@@ -7,30 +7,69 @@ import Loader from './components/Loader/Loader';
 import './App.css';
 import Scroll from './components/ScrollToTop';
 
-const Login = React.lazy(()=>import("./components/Login"));
-const Register = React.lazy(()=>import('./components/PatientRegister'));
-const DoctorRegister = React.lazy(()=>import('./components/DoctorRegister'));
-const Appointments = React.lazy(()=>import('./components/Appointments'));
-const Emergency = React.lazy(()=>import('./components/Emergency'));
-const Header = React.lazy(()=>import('./components/HeaderComponent'));
-const Forget = React.lazy(()=>import('./components/Forget'));
-const Covidtest = React.lazy(()=>import('./components/Covidtest'));
-const Covidpatient = React.lazy(()=>import('./components/CovidPatient'));
-
+const Login = React.lazy(() => {
+  return new Promise(resolve => {
+    setTimeout(() => resolve(import('./components/Login')), 3000);
+  });
+});
+const Register = React.lazy(() => {
+  return new Promise(resolve => {
+    setTimeout(() => resolve(import('./components/PatientRegister')), 3000);
+  });
+});
+const DoctorRegister = React.lazy(() => {
+  return new Promise(resolve => {
+    setTimeout(() => resolve(import('./components/DoctorRegister')), 3000);
+  });
+});
+const Appointments = React.lazy(() => {
+    return new Promise(resolve => {
+      setTimeout(() => resolve(import('./components/Appointments')), 3000);
+    });
+  });
+const Emergency = React.lazy(() => {
+    return new Promise(resolve => {
+      setTimeout(() => resolve(import('./components/Emergency')), 3000);
+    });
+  });
+const Header = React.lazy(() => {
+    return new Promise(resolve => {
+      setTimeout(() => resolve(import('./components/HeaderComponent')), 3000);
+    });
+  });
+const Forget = React.lazy(() => {
+    return new Promise(resolve => {
+      setTimeout(() => resolve(import('./components/Forget')), 3000);
+    });
+  });
+const Covidtest = React.lazy(() => {
+    return new Promise(resolve => {
+      setTimeout(() => resolve(import('./components/Covidtest')), 3000);
+    });
+  });
+const Covidpatient = React.lazy(() => {
+    return new Promise(resolve => {
+      setTimeout(() => resolve(import('./components/CovidPatient')), 3000);
+    });
+  });
 
 function App() {
-
+  const[dark,setMode] = useState(false)
     return (
-      <div className="App">
+      <div className={dark ? "App dark-mode": "App"}>
         <NavbarItem />
-        {/* <Login /> */}
-        {/* <Header /> */}
-        {/* <DoctorRegister /> */}
-        {/* <Register /> */}
+        <div className="nav2" >
+                        <label className="switch" >
+                        <input type="checkbox" onChange={()=>setMode(!dark)}/>
+                        <span className="slider round"></span>
+                        </label>
+                        </div>
         <BrowserRouter>
+
         <Suspense fallback={<Loader />}>
           <Switch>
             <Route
+
               path="/Login"
               component={Login}
             /> 
@@ -46,7 +85,6 @@ function App() {
               path="/patients"
               component={Register}
             />
-            
             <Route 
             exact path="/covidpatient"
             component={Covidpatient}
@@ -65,6 +103,7 @@ function App() {
             />
             <Route
               path=""
+
               component={Header}
             />       
 
@@ -75,8 +114,6 @@ function App() {
         <Footer />
       </div>
     );
-
-
 }
 
 export default App
