@@ -6,8 +6,10 @@ import hospital from '../assets/hospital.jpg';
 import {Multiselect} from 'multiselect-react-dropdown';
 import { GoogleLogin } from 'react-google-login';
 
+
 export default function CovidPatient() {
 
+    const [dark, setMode] = useState(false)
     const data1 = [
         
         {
@@ -57,7 +59,13 @@ const data2 = [
     return (
         <div className="container">
                 <div className="row  justify-content-center ">
-                    <Card className="mt-5 col-12 col-md-6 items">
+                <Card className={dark ? "mt-5 col-12 col-md-6 items dark-mode": "mt-5 col-12 col-md-6 items"}>
+                    <div className="nav">
+                            <label className="switch">
+                                <input type="checkbox" onChange={()=>setMode(!dark)}/>
+                                <span className="slider round"></span>
+                            </label>
+                        </div>
                         <h1 className="title mt-5">Register as Covid-19 patients</h1>
 
                         <CardBody>
@@ -83,10 +91,12 @@ const data2 = [
                                 <div className="selection">
                                     <Multiselect  options={selections} displayValue="label" placeholder="Any other Symptoms"/>
                                 </div>
+
                                 <br/>
                                 <button className="red ripple">Submit</button><br></br>
                                 <p >OR</p>
                                 <GoogleLogin buttonText="Sign in with Google" />
+
                             </form>
                             <p className="linkitem mt-3">Have an account ? <a href="Login">Login</a> </p>
                         </CardBody>
