@@ -1,13 +1,19 @@
+
+
 import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
+import './Emergency.css';
+import "./PatientRegister";
+
 import './Emergency.css'
+
 import axios from 'axios';
 import Ambulance from '../assets/ambulance.svg'
 import { Card, CardBody } from 'reactstrap'
 import { backend_url } from '../config';
 
 const Emergency = () => {
-
+  const[dark,setMode] = useState(false)
     const history = useHistory()
     const [email, setEmail] = useState("")
     const [name, setName] = useState("")
@@ -24,7 +30,7 @@ const Emergency = () => {
                 location,
             })
             .then(res => {
-                if (res.status == 200) {
+                if (res.status === 200) {
                     alert('Ambulance will be dispatched soon!')
                     history.push('/')
                 }
@@ -35,7 +41,13 @@ const Emergency = () => {
     return (
         <div className="container">
             <div className="row  justify-content-center ">
-                <Card className="mt-5 col-12 col-md-6 items">
+                <Card className={dark ? "mt-5 col-12 col-md-6 items dark-mode": "mt-5 col-12 col-md-6 items"}>
+                <div className="nav">
+                            <label className="switch">
+                                <input type="checkbox" onChange={()=>setMode(!dark)}/>
+                                <span className="slider round"></span>
+                            </label>
+                        </div>
                     <h1 className="title mt-5">Book an Ambulance</h1>
                     <CardBody>
                         <img
@@ -72,7 +84,7 @@ const Emergency = () => {
                                 value={location}
                                 onChange={(e) => setLocation(e.target.value)}
                             />
-                            <button className="red ripple" type='submit'>
+                            <button className="button" type='submit'>
                                 Call Ambulance{' '}
                             </button>
                         </form>
@@ -81,6 +93,7 @@ const Emergency = () => {
             </div>
         </div>
     )
+
 
 }
 
