@@ -12,13 +12,13 @@ import { Collapse } from 'reactstrap'
 import './Navbar.css'
 
 class NavbarItem extends Component {
+    
     constructor(props) {
         super(props)
         this.state = {
             isNavOpen: false,
         }
-
-        this.toggleNav = this.toggleNav.bind(this)
+            this.toggleNav = this.toggleNav.bind(this)
     }
 
     toggleNav() {
@@ -31,36 +31,31 @@ class NavbarItem extends Component {
         return (
             <div className="nav-outer">
                 <React.Fragment>
-                    <Navbar dark color="#d81b60" expand="md" className="navbar">
+                    <Navbar dark color="#d81b60" expand="md" className="navbar" >
+                        <div className="leftside">
+                                <NavbarBrand className="ml-1 head" href="/">lifeMEDeasy</NavbarBrand>
+                           
+                        </div>
                         <div className="container">
-                            <NavbarBrand className="ml-1" href="/">
-                                lifeMEDeasy
-                            </NavbarBrand>
-                            <NavbarToggler onClick={this.toggleNav} />
+                            <NavbarToggler onClick={this.toggleNav} className="toggle" />
+                            <Collapse isOpen={this.state.isNavOpen} navbar >
+                             
+                              
+                                <Nav className="ml-auto Navitem rightside" navbar>
+                                    <NavItem active>
+                                        <NavLink  href="Doctors" className="links">Doctors </NavLink>
+                                    </NavItem>
+                                    <NavItem active>
+                                        <NavLink href="patients" className="links"><span className="fa fa-info fa-lg"></span>Patients </NavLink>
+                                    </NavItem>
+                                    <NavItem active>
+                                        <NavLink href="/covidpatient" className="links" ><span className="fa fa-info fa-lg"></span> Covid-19 Patients  </NavLink>
+                                    </NavItem>
+                                    <NavItem active>
+                                        <NavLink href="/covidtest" className="links" ><span className="fa fa-info fa-lg"></span> Covid-19 Tests  </NavLink>
+                                    </NavItem>
+                                    <NavItem active>
 
-
-                            <Collapse isOpen={this.state.isNavOpen} navbar>
-                                <Nav navbar className="Navitem"></Nav>
-
-                                <Nav className="ml-auto Navitem" navbar>
-                                    <NavItem active>
-                                        <NavLink href="Doctors">
-                                            Doctors{' '}
-                                        </NavLink>
-                                    </NavItem>
-                                    <NavItem active>
-                                        <NavLink href="patients">
-                                            <span className="fa fa-info fa-lg"></span>
-                                            Patients{' '}
-                                        </NavLink>
-                                    </NavItem>
-                                    <NavItem active>
-                                        <NavLink href="/covidpatient"  ><span className="fa fa-info fa-lg"></span> Covid-19 Patients  </NavLink>
-                                    </NavItem>
-                                    <NavItem active>
-                                        <NavLink href="/covidtest"  ><span className="fa fa-info fa-lg"></span> Covid-19 Tests  </NavLink>
-                                    </NavItem>
-                                    <NavItem active>
                                         <NavLink href="/emergency"  ><span className="fa fa-info fa-lg"></span> Emergencies</NavLink>
                                     </NavItem>
                                     <NavItem active>
@@ -68,17 +63,14 @@ class NavbarItem extends Component {
                                             <span className="fa fa-info fa-lg"></span>{' '}
                                             Appointments
                                         </NavLink>
+
                                     </NavItem>
-                                    
-                                    
+
                                     <NavItem active>
-                                        <NavLink href="/">
-                                            <span className="fa fa-info fa-lg"></span>{' '}
-                                            Let&#39;s Talk{' '}
-                                        </NavLink>
+                                        <NavLink href="/" className="links" ><span className="fa fa-info fa-lg"></span> Let&apos;s Talk </NavLink>
                                     </NavItem>
                                     <NavItem active>
-                                        <NavLink href="#">
+                                        <NavLink href="#" className="links">
                                             <span className="fa fa-info fa-lg"></span>{' '}
                                             SignIn/SignUp{' '}
                                         </NavLink>
